refactor: Extract getFilteredStoreUrl() and move StoreButton
